+++ absp/src/ADQ14_FWPD.cpp
@@ -166,9 +166,9 @@ int ABCD::ADQ14_FWPD::Initialize()
         std::cout << "PD Firmware generation: " << FWPD_generation << "; ";
         std::cout << "Streaming generation: " << streaming_generation << "; ";
         std::cout << "ADQ14 Revision: {";
-        int* revision = ADQ_GetRevision(adq_cu_ptr, adq_num);
+        uint32_t* revision = ADQ_GetRevision(adq_cu_ptr, adq_num);
         for (int i = 0; i < 6; i++) {
-            std::cout << revision[i] << ", ";
+            std::cout << (unsigned int)revision[i] << ", ";
         }
         std::cout << "}; ";
         std::cout << std::endl;
